Initial commit - Add Refactored train.py
Refactored original code across train.py, models.py and utils.py.
Goal is to design a framework to quickly model the decoder, beta schedule
and degradation process used for the diffusion model. Chose to design the
code base so that input to the train script is just a config for reproducability.

Train script outputs a single runs dir which contains all the relevant output of a job.

Use weights and biases for logging.

Have yet to implement plotting logic to visualise unconditional and conditional generation.