BugFix - Fixed Unconditional Sampling Method for FashionMNIST.
Now using the test set for unconditional sampling, also correctly sets
original eps at the start and avoids over writing it with a fainter version.
As was the case when original eps was set inside restore.

Added a test arg to degrade to set which fashion mnist case it uses to degrade
the mnist case with.

Added a fine tuned unconditional sampling method to the class which appropriately
sets eps_original for restoration.